Add Test cases for multiple configuration element
